rpc 注册中心、注册中心优化

+++ example-common/src/main/java/com/yuxin/example/common/service/UserService.java
@@ -20,6 +20,6 @@ public interface UserService {
      * 新方法 - 获取数字
      */
     default short getNumber() {
-        return 1;
+        return (short) 1;
     }
 }
